+++ src/utils/testing/setup.py
@@ -3,7 +3,7 @@ from typing import Optional, List
 from algosdk.v2client.algod import AlgodClient
 from algosdk.kmd import KMDClient
 
-from utils.account import Account
+from src.utils.account import Account
 
 ALGOD_ADDRESS = "http://localhost:4001"
 ALGOD_TOKEN = "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
