+++ tox.ini
@@ -14,9 +14,9 @@ python =
 passenv = *
 deps =
     pytest
+    -rrequirements.txt
     pytest-cov
     pytest-dotenv
-    pytest-asyncio
     pytest-xdist
 commands =
     python -m pytest -n 2 --cov --cov-report xml
